pagination

+++ app/Http/Controllers/ReservationController.php
@@ -100,9 +100,12 @@ class ReservationController extends Controller
      */
     public function destroy(string $id)
     {
+        if (! Gate::allows('destroy-reservation', Reservation::all()->where('id', $id)->first())) {
+            return redirect('/error')->with('message',
+            'У вас нет разрешения на удаление бронирования номер'. $id);
+        }
         Reservation::destroy($id);
-        return redirect('/reservation')->withErrors([
-            'success' => 'Запись '. $id . ' успешно удалена!'
-        ]);
+        return redirect('/reservation');
+
     }
 }
